Rename to helper.covers.ts for consistency
