+++ components/pages/workPageLayout/WorkPageHeader.tsx
@@ -17,7 +17,7 @@ import { Badge } from "@/components/shared/badge/Badge"
 import { CoverPicture } from "@/components/shared/coverPicture/CoverPicture"
 import SlideSelect, { SlideSelectOption } from "@/components/shared/slideSelect/SlideSelect"
 import { WorkFullWorkPageFragment } from "@/lib/graphql/generated/fbi/graphql"
-import { getCoverUrls, getLowResCoverUrl } from "@/lib/helpers/covers"
+import { getCoverUrls, getLowResCoverUrl } from "@/lib/helpers/helper.covers"
 import { displayCreators } from "@/lib/helpers/helper.creators"
 import { getIsbnsFromManifestation } from "@/lib/helpers/ids"
 import { useGetCoverCollection } from "@/lib/rest/cover-service-api/generated/cover-service"
