+++ lib/helpers/helper.covers.ts
